feat(script): make EVAL & FCALL exclusive if lua-strict-key-accessing is disabled (#3155)
Accessing undeclared keys may lead to unexpected behavior in both
Kvrocks and Redis.

Specifically in Kvrocks, undeclared keys will not be protected by key
locks, which may face atomic issues of rocksdb IO.

If `lua-strict-key-accessing` is enabled, we can garantee that there is
no write on undeclared keys, so EVAL & FCALL in multiple threads should
be safe.

But for disabled `lua-strict-key-accessing`, it is quite dangerous if
users use undeclared keys in scripting. So we add `exclusive` flag to
EVAL, EVALSHA and FCALL commands in such case.

NOTE that *_RO commands are not affected.