feat(ts): Add `TS.INCRBY` and `TS.DECRBY` command  (#3171)
Part of #3048 

In the current implementation, `TS.INCRBY` defaults to using `0` as the
timestamp when `TIMESTAMP` is not specified, which differs from Redis'
native behavior (using the server's timestamp).
This feature could be implemented in the future, along with the `TS.ADD
key * value` syntax to support automatically using the server's
timestamp.


```
127.0.0.1:6666>  TS.INCRBY a 232 TIMESTAMP 1657811829000
(integer) 1657811829000
127.0.0.1:6666> TS.INCRBY a 157 TIMESTAMP 1657811829000
(integer) 1657811829000
127.0.0.1:6666> TS.INCRBY a 432 TIMESTAMP 1657811829000
(integer) 1657811829000
127.0.0.1:6666> ts.get a
1) 1) (integer) 1657811829000
   2) (double) 821
127.0.0.1:6666> ts.incrby a 100 TIMESTAMP 50
(error) ERR timestamp must be equal to or higher than the maximum existing timestamp
127.0.0.1:6666> 
```