fix(replication): send ack on ping even if no data written (#3220)
In the old code, we only ack when there is data written and
`sendReplConfAck` would only send ack if not ack was sent in the last 1
sec. This works fine for continuous traffic. However, if write stops and
the last write did not trigger an ack due to 1s rate limit, no more ack
would be sent.

If user runs `INFO replication` on master, the user would see a
replication lag that never get catch up. In this PR we send ack back
whenever replication sees ping, because ping is only sent when master
has no data, this should not add extra load.

---------

Co-authored-by: Twice <twice@apache.org>