feat(replication): add replication-no-slowdown config (#3219)
We need to set `rocksdb.write_options.no_slowdown` to yes in production
because if write stall happens, it would block the event loop and
KVRocks would not be able to serve any other requests including `INFO`.
This means when write stall happens, we would see the cluster frozen and
failover happens unnecessarily.

However, a side effect of `rocksdb.write_options.no_slowdown` is it
would cause replication failure when write stall happens. The replica
would continuously tries to reconnect and restart the process, adding
extra load on master.

The solution is to add a separate config to fine control the behavior.

---------

Co-authored-by: Twice <twice@apache.org>