ci: update clang-format from 14 to 18 (#3221)
Recently, we’ve seen many PRs fail the CI check because they were
formatted with a newer version of clang-format, which differs from the
one used in CI (clang-format-14). Clang-format 14 is a relatively old
version, especially on newer Linux distributions.

In this PR, we switch to using clang-format 18. We didn’t choose a
higher version because clang-format 18 is the latest one available in
Ubuntu 24.04’s official apt repository. Although LLVM provides newer
versions through its own apt source, we’re upgrading to 18 for now to
make it easier for developers to install and use.